BigData/requirements.txt
pyspark==3.5.0
kafka-python==2.3.1
pandas>=1.5.0
pyarrow>=12.0.0
